i2c: rcar: skip DMA if buffer is not safe
This HW is prone to races, so it needs to setup new messages in irq
context. That means we can't alloc bounce buffers if a message buffer is
not DMA safe. So, in that case, simply fall back to PIO.

Reviewed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
(cherry picked from commit adbd77b589d0345cfc42f1ae0b3414e62a2bf866)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>

Renamed rcar_i2c_is_dma() to rcar_i2c_is_pio() and added the same
DMA-safety test to the function. This is done in v4.14.75 commit
55d2d2fb8b04b4b7bbdbd68d5e0523841b610de2 as well.

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>